+++ example/lib/lifecycle/lifecycle_model.dart
@@ -1,0 +1,3 @@
+import 'package:mvc/base/base_model.dart';
+
+class LifecycleModel extends BaseModel {}